LibFileSystem+Userland: Rename size() to size_from_stat()
This reflects what the functions does more accurately, and allows for
adding functions to get sizes through other methods.

This also corrects the return type of said function, as size_t may only
hold sizes up to 4GB on 32-bit platforms.

+++ Userland/Services/WebServer/Client.h
@@ -28,7 +28,7 @@ private:
 
     struct ContentInfo {
         String type;
-        size_t length {};
+        u64 length {};
     };
 
     ErrorOr<void, WrappedError> on_ready_to_read();
